requirements.txt
asgiref==3.5.2
Django==4.0.1
django-environ==0.9.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
djangorestframework==3.14.0
